+++ src/main/java/in/koreatech/koin/admin/user/dto/AdminStudentsResponse.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Page;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
 public record AdminStudentsResponse(
