+++ src/main/java/in/koreatech/koin/admin/user/service/AdminUserService.java
@@ -44,10 +44,10 @@ import in.koreatech.koin.domain.owner.model.OwnerIncludingShop;
 import in.koreatech.koin.domain.owner.model.OwnerShop;
 import in.koreatech.koin.domain.shop.model.shop.Shop;
 import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
-import in.koreatech.koin.domain.user.exception.StudentDepartmentNotValidException;
+import in.koreatech.koin.domain.student.exception.StudentDepartmentNotValidException;
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
-import in.koreatech.koin.domain.user.model.Student;
-import in.koreatech.koin.domain.user.model.StudentDepartment;
+import in.koreatech.koin.domain.student.model.Student;
+import in.koreatech.koin.domain.student.model.StudentDepartment;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.domain.user.model.UserGender;
 import in.koreatech.koin.domain.user.model.UserToken;
