+++ src/main/java/in/koreatech/koin/domain/shop/model/review/ShopReview.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import in.koreatech.koin.domain.shop.model.shop.Shop;
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.global.domain.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
