+++ src/main/java/in/koreatech/koin/domain/shop/model/review/ShopReviewReport.java
@@ -3,7 +3,7 @@ package in.koreatech.koin.domain.shop.model.review;
 import static jakarta.persistence.FetchType.LAZY;
 import static lombok.AccessLevel.PROTECTED;
 
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.global.domain.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
