+++ src/main/java/in/koreatech/koin/domain/shop/service/ShopReviewService.java
@@ -37,8 +37,8 @@ import in.koreatech.koin.domain.shop.repository.review.ShopReviewMenuRepository;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewReportCategoryRepository;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewReportRepository;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewRepository;
-import in.koreatech.koin.domain.user.model.Student;
-import in.koreatech.koin.domain.user.repository.StudentRepository;
+import in.koreatech.koin.domain.student.model.Student;
+import in.koreatech.koin.domain.student.repository.StudentRepository;
 import in.koreatech.koin.global.auth.exception.AuthenticationException;
 import in.koreatech.koin.global.model.Criteria;
 import jakarta.persistence.EntityManager;
