+++ src/main/java/in/koreatech/koin/domain/student/model/StudentDepartment.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.user.model;
+package in.koreatech.koin.domain.student.model;
 
 import lombok.Getter;
