+++ src/main/java/in/koreatech/koin/domain/student/repository/StudentRepository.java
@@ -1,11 +1,11 @@
-package in.koreatech.koin.domain.user.repository;
+package in.koreatech.koin.domain.student.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.user.exception.UserNotFoundException;
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 
 public interface StudentRepository extends Repository<Student, Integer> {
