+++ src/main/java/in/koreatech/koin/domain/student/service/StudentService.java
@@ -1,12 +1,17 @@
-package in.koreatech.koin.domain.user.service;
+package in.koreatech.koin.domain.student.service;
 
 import java.time.Clock;
 import java.util.Optional;
 import java.util.UUID;
 
+import in.koreatech.koin.domain.student.model.Student;
+import in.koreatech.koin.domain.student.model.StudentDepartment;
+import in.koreatech.koin.domain.student.model.StudentEmailRequestEvent;
+import in.koreatech.koin.domain.student.model.StudentRegisterEvent;
+import in.koreatech.koin.domain.user.dto.UserPasswordChangeRequest;
 import in.koreatech.koin.domain.user.model.*;
-import in.koreatech.koin.domain.user.model.redis.StudentTemporaryStatus;
-import in.koreatech.koin.domain.user.repository.StudentRedisRepository;
+import in.koreatech.koin.domain.student.model.redis.StudentTemporaryStatus;
+import in.koreatech.koin.domain.student.repository.StudentRedisRepository;
 
 import org.joda.time.LocalDateTime;
 import org.springframework.context.ApplicationEventPublisher;
@@ -17,17 +22,17 @@ import org.springframework.web.servlet.ModelAndView;
 
 import in.koreatech.koin.domain.user.dto.AuthTokenRequest;
 import in.koreatech.koin.domain.user.dto.FindPasswordRequest;
-import in.koreatech.koin.domain.user.dto.StudentLoginRequest;
-import in.koreatech.koin.domain.user.dto.StudentLoginResponse;
-import in.koreatech.koin.domain.user.dto.StudentRegisterRequest;
-import in.koreatech.koin.domain.user.dto.StudentResponse;
-import in.koreatech.koin.domain.user.dto.StudentUpdateRequest;
-import in.koreatech.koin.domain.user.dto.StudentUpdateResponse;
-import in.koreatech.koin.domain.user.dto.UserPasswordChangeRequest;
+import in.koreatech.koin.domain.student.dto.StudentLoginRequest;
+import in.koreatech.koin.domain.student.dto.StudentLoginResponse;
+import in.koreatech.koin.domain.student.dto.StudentRegisterRequest;
+import in.koreatech.koin.domain.student.dto.StudentResponse;
+import in.koreatech.koin.domain.student.dto.StudentUpdateRequest;
+import in.koreatech.koin.domain.student.dto.StudentUpdateResponse;
+import in.koreatech.koin.domain.user.dto.UserPasswordChangeSubmitRequest;
 import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
-import in.koreatech.koin.domain.user.exception.StudentDepartmentNotValidException;
-import in.koreatech.koin.domain.user.exception.StudentNumberNotValidException;
-import in.koreatech.koin.domain.user.repository.StudentRepository;
+import in.koreatech.koin.domain.student.exception.StudentDepartmentNotValidException;
+import in.koreatech.koin.domain.student.exception.StudentNumberNotValidException;
+import in.koreatech.koin.domain.student.repository.StudentRepository;
 import in.koreatech.koin.domain.user.repository.UserRepository;
 import in.koreatech.koin.domain.user.repository.UserTokenRepository;
 import in.koreatech.koin.global.auth.JwtProvider;
@@ -204,7 +209,14 @@ public class StudentService {
     }
 
     @Transactional
-    public void changePassword(UserPasswordChangeRequest request, String resetToken) {
+    public void changePassword(Integer userId, UserPasswordChangeRequest request) {
+        Student student = studentRepository.getById(userId);
+        User user = student.getUser();
+        user.updateStudentPassword(passwordEncoder, request.password());
+    }
+
+    @Transactional
+    public void changePasswordSubmit(UserPasswordChangeSubmitRequest request, String resetToken) {
         User authedUser = userRepository.getByResetToken(resetToken);
         authedUser.validateResetToken();
         authedUser.updatePassword(passwordEncoder, request.password());
