+++ src/main/java/in/koreatech/koin/domain/user/dto/UserPasswordChangeSubmitRequest.java
@@ -2,14 +2,14 @@ package in.koreatech.koin.domain.user.dto;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record UserPasswordChangeRequest(
+@JsonNaming(SnakeCaseStrategy.class)
+public record UserPasswordChangeSubmitRequest(
     @Schema(description = "변경할 비밀번호 (SHA 256 해싱된 값)", example = "password", requiredMode = REQUIRED)
     @NotBlank(message = "비밀번호를 입력해주세요.")
     String password
