+++ src/test/java/in/koreatech/koin/acceptance/ShopReviewApiTest.java
@@ -20,7 +20,6 @@ import org.springframework.transaction.support.TransactionTemplate;
 
 import in.koreatech.koin.AcceptanceTest;
 import in.koreatech.koin.domain.owner.model.Owner;
-import in.koreatech.koin.domain.shop.model.review.ReportStatus;
 import in.koreatech.koin.domain.shop.model.review.ShopReview;
 import in.koreatech.koin.domain.shop.model.review.ShopReviewReport;
 import in.koreatech.koin.domain.shop.model.review.ShopReviewReportCategory;
@@ -28,7 +27,7 @@ import in.koreatech.koin.domain.shop.model.shop.Shop;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewReportCategoryRepository;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewReportRepository;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewRepository;
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.fixture.ShopFixture;
 import in.koreatech.koin.fixture.ShopReviewFixture;
 import in.koreatech.koin.fixture.ShopReviewReportCategoryFixture;
