+++ src/test/java/in/koreatech/koin/fixture/ShopReviewReportFixture.java
@@ -6,7 +6,7 @@ import in.koreatech.koin.domain.shop.model.review.ReportStatus;
 import in.koreatech.koin.domain.shop.model.review.ShopReview;
 import in.koreatech.koin.domain.shop.model.review.ShopReviewReport;
 import in.koreatech.koin.domain.shop.repository.review.ShopReviewReportRepository;
-import in.koreatech.koin.domain.user.model.Student;
+import in.koreatech.koin.domain.student.model.Student;
 
 @Component
 @SuppressWarnings("NonAsciiCharacters")
